requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
playwright
python-dotenv
pydantic
requests
nfl_data_py
scikit-learn
